firstproject/requirements.txt
appdirs==1.4.0
Django==1.10.5
django-webpack-loader==0.4.1
packaging==16.8
pyparsing==2.1.10
six==1.10.0
django-bower==5.2.0
django-scheduler==0.8
